pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigma-detection-pipeline"
version = "0.1.0"
description = "A detection engineering framework for Sigma rules"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
keywords = ["sigma", "detection", "security", "siem", "threat-hunting"]
dependencies = [
    "pyyaml>=6.0",
    "pysigma>=0.11.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
backends = [
    "pysigma-backend-kusto>=0.4.0",
    "pysigma-backend-splunk>=1.0.0",
]

[project.scripts]
sigma-pipeline = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/sigma-detection-pipeline"
Repository = "https://github.com/yourusername/sigma-detection-pipeline"
Issues = "https://github.com/yourusername/sigma-detection-pipeline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
